.characters leaderboard and summary, and working code redemption commands

+++ src/main/java/com/tb24/discordbot/Session.kt
@@ -39,7 +39,7 @@ class Session @JvmOverloads constructor(val client: DiscordBot, val id: String, 
 	@Throws(HttpException::class, IOException::class)
 	fun login(source: CommandSourceStack?, grantType: GrantType, fields: Map<String, String>, auth: EAuthClient = EAuthClient.FORTNITE_IOS_GAME_CLIENT, sendMessages: Boolean = true): Int {
 		if (source != null) {
-			if (grantType != GrantType.device_auth && grantType != GrantType.device_code && source.message.isFromGuild) {
+			if (grantType != GrantType.device_auth && grantType != GrantType.device_code && source.message.isFromGuild && sendMessages) {
 				source.message.delete().queue()
 			}
 			if (api.userToken != null) {
